Lesson 52 - Adding Visual Feedback to the Buttons
This moves the `Pressable` component inside the `View` to properly show the Android ripple effect. Doing this also needs to split the container into outer and inner containers. Also, the style of the `Pressable` component is checking if it is 'pressed' and if so it shows an array of styles, otherwise it just shows one style.