53c - My refactoring
It didn't make sense to me to add another view and styling within the main Start Screen for the buttons when the PrimaryButton component already had this view. So, I changed the code in the Start Screen and added the `flex: 1` to the PrimaryButton component. It works the same and less code.