Lesson 59 - Switching Screens on Confirmed Number
The App now has a state to determine which screen to show. If the `userNumber` is not confirmed, then it shows the start screen, otherwise it shows the game screen. Also, now the function to set the user number is passed to the start game screen. When the number is confirmed on the start screen it is passed then to the App component which then shows the game screen.