65b - Fixing an infinite loop
This now checks if the user lies and says the number is higher or lower than the guess when it is not. Before if they did lie, then it would get to a point where the min and max values were the same and the random generator would enter an infinite loop trying to guess a number.