Lesson 66 - Coding the End of Game Logic
This commit shows the use of State for end of game, sets it to default as true, and switch to false once a number is chosen and back to true when the number is guessed. The function is passed to the GameScreen as a prop.

Game screen uses `useEffect` to run a code snippet whenever the dependencies of the `useEffect` change. So, in this case whenever `currentGuess`, `userNumber` and/or `onGameOver` are changed. Also, the initial guess numbers need to be hard coded as the code is evaluated whenever the GameScreen refreshes and when the correct number is found an infinite loop is triggered. Now that they are hard coded, this does not occur.

For now, the Game Over screen just outputs that the game is over.