My own refactoring
To test out my knowledge, I thought that I saw a way to refactor and limit code. We shall see if this breaks something in the future that will be taught as a previous refactor for the '52c' commit showed.

In this refactor, I didn't see the need to set the 'Game is Over' as true to begin with. If it is set as false, because the game hasn't even begun yet, then it wouldn't need to be set to false in the `pickedNumberHandler` and only the `gameIsOver` would need to be checked in line 31(previously 32).

I checked and ran this code and it works the same. Now with less code to have to maintain and I think it reads better.