Lesson 68 - Quasi-Cascading Styles
This method allows for passing styles from a parent component like 'InstructionText', to a child component like 'Text' within the 'InstructionText'. This method also shows the merging of styles via an array. They are applied left to right, so anything defined in a later style within the array may be used to overwrite what was stated previously.

In this example I could have put the 'style' sent via props to the `InstructionText` component at the beginning of the array to insure that the styles I implement within the component take precedence; or I could have it at the end, which allows the props to overwrite the standard styles.

I could also do it both ways as seen in this example, where I could have a style within the `InstructionText` component titled 'requiredStyle' and another titled 'defaultStyle'. Then I could pass via a prop the styles I want to overwrite. So, looking at this code on line 6 of the `InstructionText`, you can see that the requiredStyle is set last so it will take precedence, and the default style will be able to be overwriten by the prop style (ie the marginBottom attribute).