Lesson 85 - KeyboardAvoidingView
From time to time the keyboard of the device may block parts of the screen when activated. `KeyboardAvoidingView` takes care of this. But, you need to wrap all content with this component, then all of that with a `ScrollView` and set both the `KeyboardAvoidingView` and `ScrollView` with a `flex: 1` property.

Also, `KeyboardAvoidingView` has a `behavior` prop that needs to be set to make it work.

`KeyboardAvoidingView` also allows the iOS keyboard to be closed by clicking outside the keyboard.