88b - Another way of writing OS specific code
The components, colors, etc can all have the platform added to the file name to setup differences per platform. For this example, it is a bit overkill as it is only changing whether there is a border or not and so the other code is best. But, it is good to know that you can call `import Title from "../components/ui/Title";` to bring in the Title component and under the hood React Native will choose the corresponding OS file on creation of the app.